jsHistoryBack, jsReplace, doDelete confirm, doModify 구조변경, Rq,interceptor 구현중

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -45,7 +45,7 @@
 				
 				<c:if test="${loginedMemberId == article.memberId }">
 					<a class="btn-text-link" href="modify?id=${article.id}">수정</a>
-					<a class="btn-text-link" href="doDelete?id=${article.id}">삭제</a>
+					<a class="btn-text-link" href="doDelete?id=${article.id}" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
 				</c:if>
 			</div>
 		</div>
